# [libc++][format][4/7] Improves std::format_to performance.