This changes the __output_buffer users for format_to to the new structure. It also uses the new allocating buffer to improve performance.

write_double_comparison:

Before
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_sprintf                                           197 ns          197 ns      3537000
BM_to_string                                         219 ns          219 ns      3200000
BM_to_chars                                         42.3 ns         42.3 ns     16490000
BM_to_chars_as_string                               48.1 ns         48.1 ns     14508000
BM_format                                            167 ns          167 ns      4207000
BM_format_to_back_inserter<std::string>              179 ns          179 ns      3914000
BM_format_to_back_inserter<std::vector<char>>        216 ns          216 ns      3234000
BM_format_to_back_inserter<std::list<char>>          752 ns          752 ns       932000
BM_format_to_iterator/<std::array>                   165 ns          165 ns      4257000
BM_format_to_iterator/<std::string>                  165 ns          165 ns      4251000
BM_format_to_iterator/<std::vector>                  165 ns          165 ns      4254000

After
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_sprintf                                           197 ns          197 ns      3548000
BM_to_string                                         219 ns          219 ns      3203000
BM_to_chars                                         42.3 ns         42.3 ns     16540000
BM_to_chars_as_string                               48.2 ns         48.2 ns     14532000
BM_format                                            167 ns          167 ns      4190000
BM_format_to_back_inserter<std::string>              176 ns          176 ns      3959000
BM_format_to_back_inserter<std::vector<char>>        211 ns          211 ns      3312000
BM_format_to_back_inserter<std::list<char>>          752 ns          752 ns       934000
BM_format_to_iterator/<std::array>                   162 ns          162 ns      4313000
BM_format_to_iterator/<std::string>                  162 ns          162 ns      4307000
BM_format_to_iterator/<std::vector>                  162 ns          162 ns      4308000

Comparison
Benchmark                                                       Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------------------------
BM_sprintf                                                   +0.0003         +0.0002           197           197           197           197
BM_to_string                                                 -0.0004         -0.0004           219           219           219           219
BM_to_chars                                                  +0.0003         +0.0003            42            42            42            42
BM_to_chars_as_string                                        +0.0004         +0.0004            48            48            48            48
BM_format                                                    +0.0016         +0.0016           167           167           167           167
BM_format_to_back_inserter<std::string>                      -0.0128         -0.0127           179           176           179           176
BM_format_to_back_inserter<std::vector<char>>                -0.0216         -0.0216           216           211           216           211
BM_format_to_back_inserter<std::list<char>>                  -0.0008         -0.0008           752           752           752           752
BM_format_to_iterator/<std::array>                           -0.0137         -0.0137           165           162           165           162
BM_format_to_iterator/<std::string>                          -0.0154         -0.0153           165           162           165           162
BM_format_to_iterator/<std::vector>                          -0.0138         -0.0138           165           162           165           162
OVERALL_GEOMEAN                                              -0.0069         -0.0069             0             0             0             0


write_int_comparison:

Before
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_sprintf                                          80.3 ns         80.1 ns      8703000
BM_to_string                                        15.0 ns         14.9 ns     46849000
BM_to_chars                                         4.94 ns         4.92 ns    139052000
BM_to_chars_as_string                               15.6 ns         15.6 ns     44945000
BM_format                                           62.1 ns         61.9 ns     11302000
BM_format_to_back_inserter<std::string>             70.3 ns         70.1 ns      9984000
BM_format_to_back_inserter<std::vector<char>>       92.9 ns         92.7 ns      7520000
BM_format_to_back_inserter<std::list<char>>          239 ns          239 ns      2926000
BM_format_to_iterator/<std::array>                  60.6 ns         60.4 ns     11582000
BM_format_to_iterator/<std::string>                 60.2 ns         60.0 ns     11671000
BM_format_to_iterator/<std::vector>                 60.1 ns         60.0 ns     11691000

After
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_sprintf                                          81.3 ns         81.2 ns      8546000
BM_to_string                                        14.9 ns         14.9 ns     46809000
BM_to_chars                                         4.93 ns         4.93 ns    138278000
BM_to_chars_as_string                               15.4 ns         15.4 ns     45504000
BM_format                                           62.0 ns         61.9 ns     11330000
BM_format_to_back_inserter<std::string>             69.0 ns         68.9 ns     10164000
BM_format_to_back_inserter<std::vector<char>>       90.4 ns         90.3 ns      7736000
BM_format_to_back_inserter<std::list<char>>          242 ns          242 ns      2891000
BM_format_to_iterator/<std::array>                  58.6 ns         58.6 ns     11952000
BM_format_to_iterator/<std::string>                 58.2 ns         58.1 ns     12052000
BM_format_to_iterator/<std::vector>                 58.4 ns         58.3 ns     11983000

Comparison
Benchmark                                                       Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------------------------
BM_sprintf                                                   +0.0116         +0.0132            80            81            80            81
BM_to_string                                                 -0.0050         -0.0036            15            15            15            15
BM_to_chars                                                  -0.0005         +0.0010             5             5             5             5
BM_to_chars_as_string                                        -0.0168         -0.0149            16            15            16            15
BM_format                                                    -0.0015         +0.0001            62            62            62            62
BM_format_to_back_inserter<std::string>                      -0.0186         -0.0172            70            69            70            69
BM_format_to_back_inserter<std::vector<char>>                -0.0271         -0.0256            93            90            93            90
BM_format_to_back_inserter<std::list<char>>                  +0.0104         +0.0119           239           242           239           242
BM_format_to_iterator/<std::array>                           -0.0322         -0.0304            61            59            60            59
BM_format_to_iterator/<std::string>                          -0.0330         -0.0317            60            58            60            58
BM_format_to_iterator/<std::vector>                          -0.0285         -0.0270            60            58            60            58
OVERALL_GEOMEAN                                              -0.0130         -0.0114             0             0             0             0


write_string_comparison:
Comparison
Benchmark                                                                              Time             CPU      Time Old      Time New       CPU Old       CPU New
-------------------------------------------------------------------------------------------------------------------------------------------------------------------
BM_sprintf/C string len = 6                                                         -0.0037         -0.0051             4             4             4             4
BM_format/C string len = 6                                                          -0.0022         -0.0037            50            49            50            49
BM_format_to_back_inserter<std::string>/C string len = 6                            -0.0136         -0.0151            57            56            57            56
BM_format_to_back_inserter<std::vector<char>>/C string len = 6                      +0.0408         +0.0392            73            76            73            76
BM_format_to_back_inserter<std::deque<char>>/C string len = 6                       +0.0115         +0.0099           156           158           156           157
BM_format_to_back_inserter<std::list<char>>/C string len = 6                        +0.0325         +0.0307           136           140           136           140
BM_format_to_iterator/<std::array> C string len = 6                                 -0.0153         -0.0167            47            46            47            46
BM_format_to_iterator/<std::string> C string len = 6                                -0.0332         -0.0346            47            46            47            46
BM_format_to_iterator/<std::vector> C string len = 6                                -0.0125         -0.0141            46            46            46            46
BM_format_to_iterator/<std::deque> C string len = 6                                 +0.0041         +0.0026            52            53            52            53
BM_format/string len = 6                                                            -0.0170         -0.0185            50            49            50            49
BM_format_to_back_inserter<std::string>/string len = 6                              -0.0092         -0.0106            56            55            56            55
BM_format_to_back_inserter<std::vector<char>>/string len = 6                        +0.0153         +0.0138            73            74            73            74
BM_format_to_back_inserter<std::deque<char>>/string len = 6                         +0.0392         +0.0371           155           161           154           160
BM_format_to_back_inserter<std::list<char>>/string len = 6                          +0.0367         +0.0351           133           138           133           138
BM_format_to_iterator/<std::array> string len = 6                                   -0.0107         -0.0125            47            46            47            46
BM_format_to_iterator/<std::string> string len = 6                                  -0.0312         -0.0326            47            46            47            46
BM_format_to_iterator/<std::vector> string len = 6                                  -0.0103         -0.0144            46            46            46            46
BM_format_to_iterator/<std::deque> string len = 6                                   -0.0042         -0.0057            53            53            53            52
BM_format/string_view len = 6                                                       -0.0191         -0.0206            49            48            49            48
BM_format_to_back_inserter<std::string>/string_view len = 6                         -0.0283         -0.0298            56            55            56            54
BM_format_to_back_inserter<std::vector<char>>/string_view len = 6                   +0.0408         +0.0389            72            75            72            75
BM_format_to_back_inserter<std::deque<char>>/string_view len = 6                    +0.0214         +0.0199           153           157           153           156
BM_format_to_back_inserter<std::list<char>>/string_view len = 6                     +0.0407         +0.0393           132           137           132           137
BM_format_to_iterator/<std::array> string_view len = 6                              -0.0235         -0.0249            46            45            46            45
BM_format_to_iterator/<std::string> string_view len = 6                             -0.0368         -0.0385            46            45            46            44
BM_format_to_iterator/<std::vector> string_view len = 6                             -0.0293         -0.0308            46            45            46            45
BM_format_to_iterator/<std::deque> string_view len = 6                              +0.0094         +0.0079            52            52            52            52
BM_sprintf/C string len = 60                                                        -0.0004         -0.0018             5             5             5             5
BM_format/C string len = 60                                                         +0.0250         +0.0232            64            65            64            65
BM_format_to_back_inserter<std::string>/C string len = 60                           -0.0027         -0.0041            75            75            75            75
BM_format_to_back_inserter<std::vector<char>>/C string len = 60                     +0.0228         +0.0233            89            91            89            91
BM_format_to_back_inserter<std::deque<char>>/C string len = 60                      +0.0303         +0.0312           285           294           285           294
BM_format_to_back_inserter<std::list<char>>/C string len = 60                       +0.0419         +0.0429          1164          1212          1163          1212
BM_format_to_iterator/<std::array> C string len = 60                                -0.0154         -0.0145            46            46            46            46
BM_format_to_iterator/<std::string> C string len = 60                               -0.0355         -0.0345            47            45            47            45
BM_format_to_iterator/<std::vector> C string len = 60                               -0.0161         -0.0152            46            46            46            46
BM_format_to_iterator/<std::deque> C string len = 60                                +0.0044         +0.0054            52            52            52            52
BM_format/string len = 60                                                           +0.0268         +0.0278            62            64            62            64
BM_format_to_back_inserter<std::string>/string len = 60                             -0.0031         -0.0022            75            74            75            74
BM_format_to_back_inserter<std::vector<char>>/string len = 60                       +0.0183         +0.0191            87            88            86            88
BM_format_to_back_inserter<std::deque<char>>/string len = 60                        +0.0567         +0.0578           279           295           279           295
BM_format_to_back_inserter<std::list<char>>/string len = 60                         +0.0386         +0.0396          1167          1212          1165          1212
BM_format_to_iterator/<std::array> string len = 60                                  -0.0094         -0.0084            47            46            47            46
BM_format_to_iterator/<std::string> string len = 60                                 -0.0268         -0.0259            47            46            47            46
BM_format_to_iterator/<std::vector> string len = 60                                 -0.0082         -0.0073            46            46            46            46
BM_format_to_iterator/<std::deque> string len = 60                                  +0.0254         +0.0264            52            54            52            54
BM_format/string_view len = 60                                                      -0.0084         -0.0075            62            62            62            62
BM_format_to_back_inserter<std::string>/string_view len = 60                        -0.0001         +0.0008            74            74            74            74
BM_format_to_back_inserter<std::vector<char>>/string_view len = 60                  +0.0145         +0.0154            86            88            86            88
BM_format_to_back_inserter<std::deque<char>>/string_view len = 60                   +0.0411         +0.0420           279           291           279           291
BM_format_to_back_inserter<std::list<char>>/string_view len = 60                    +0.0358         +0.0368          1164          1205          1163          1205
BM_format_to_iterator/<std::array> string_view len = 60                             -0.0306         -0.0297            47            45            47            45
BM_format_to_iterator/<std::string> string_view len = 60                            -0.0321         -0.0312            46            45            46            45
BM_format_to_iterator/<std::vector> string_view len = 60                            -0.0320         -0.0311            46            45            46            45
BM_format_to_iterator/<std::deque> string_view len = 60                             +0.0039         +0.0048            52            52            52            52
BM_sprintf/C string len = 6000                                                      +0.4788         +0.4802            77           114            77           114
BM_format/C string len = 6000                                                       -0.0087         -0.0078           345           342           345           342
BM_format_to_back_inserter<std::string>/C string len = 6000                         +0.0104         +0.0114          1001          1011          1000          1011
BM_format_to_back_inserter<std::vector<char>>/C string len = 6000                   +0.0195         +0.0206           972           991           971           991
BM_format_to_back_inserter<std::deque<char>>/C string len = 6000                    -0.0092         -0.0082         15129         14990         15114         14990
BM_format_to_back_inserter<std::list<char>>/C string len = 6000                     -0.0017         -0.0007        115290        115092        115167        115090
…